# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bilkis E-commerce API",
        "description": "Comprehensive API documentation for Bilkis e-commerce platform. This API provides complete functionality for managing an online store including user authentication, product catalog, shopping cart, order processing, and address management.",
        "contact": {
            "name": "Bilkis Development Team",
            "email": "bilqisshally@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        },
        {
            "url": "https://api.bilkis.com",
            "description": "Production server"
        }
    ],
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "description": "Retrieve a paginated list of all product categories",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Categories retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Category"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 45
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "description": "Create a new product category",
                "operationId": "71fcad552bb0eaba9fb191fd8d8dcab0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Category created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CreateCategoryRequest": {
                "title": "Create Category Request",
                "description": "Request body for creating a new category",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Category name (unique)",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Electronics"
                    }
                },
                "type": "object"
            },
...
            "Category": {
                "title": "Category",
                "description": "Product category model",
                "properties": {
                    "id": {
                        "description": "Category ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Category name",
                        "type": "string",
                        "example": "Electronics"
                    },
                    "slug": {
                        "description": "Category slug (auto-generated)",
                        "type": "string",
                        "example": "electronics"
                    },
                    "products_count": {
                        "description": "Number of products in this category (when loaded)",
                        "type": "integer",
                        "example": 25
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-19T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-19T10:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "description": "Validation error response",
                "properties": {
                    "code": {
                        "description": "HTTP status code",
                        "type": "integer",
                        "example": 422
                    },
                    "status": {
                        "description": "Response status",
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Validation errors by field",
                        "type": "object",
                        "example": {
                            "name": [
                                "The name field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "General error response",
                "properties": {
                    "code": {
                        "description": "HTTP status code",
                        "type": "integer",
                        "example": 404
                    },
                    "status": {
                        "description": "Response status",
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Resource not found"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication. Include the token in the Authorization header as 'Bearer {token}'",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}